+++ README.md
@@ -13,17 +13,27 @@
 
 ```bash
 azctx is a CLI tool for managing azure cli subscriptions.
-	It is a helper for the azure cli and provides a simple interface for managing subscriptions.
-	Pass a subscription name to select a specific subscription.
-	Pass - to switch to the previous subscription.
+        It is a helper for the azure cli and provides a simple interface for managing subscriptions.
+        Pass a subscription name to select a specific subscription.
+        Pass - to switch to the previous subscription.
 
 Usage:
   azctx [- / NAME] [flags]
+  azctx [command]
+
+Available Commands:
+  completion  Generate the autocompletion script for the specified shell
+  help        Help about any command
+  login       Login to Azure
+  version     Print the CLI version
 
 Flags:
-  -c, --current   Display the current active subscription
-  -h, --help      help for azctx
-  -r, --refresh   Re-Authenticate and refresh the subscriptions
+  -c, --current              Display the current active subscription
+  -t, --fetch-tenant-names   Fetch the tenant names for the subscriptions
+  -h, --help                 help for azctx
+  -r, --refresh              Re-Authenticate and refresh the subscriptions
+
+Use "azctx [command] --help" for more information about a command.
 ```
 
 -----
@@ -57,5 +67,4 @@ export AZURE_CONFIG_DIR=$HOME/.azure
 
 ## Todos
 
-* [ ] Remove dependency on `azure-cli`  
 * [ ] Implement Unit Tests
